Revert "move file from util to adapter"
This reverts commit ea64823fbdad6f233ab3d9246578693c217c1d06.
