+++ cliboa/scenario/transform/file.py
@@ -23,12 +23,12 @@ import zipfile
 
 import pandas
 
-from cliboa.adapter.file import File
 from cliboa.core.validator import EssentialParameters
 from cliboa.scenario.base import BaseStep
 from cliboa.scenario.extras import ExceptionHandler
 from cliboa.util.date import DateUtil
 from cliboa.util.exception import CliboaException, FileNotFound, InvalidParameter
+from cliboa.util.file import File
 
 
 class FileBaseTransform(BaseStep, ExceptionHandler):
